Add: Header component (#2)
